% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distro_req.R
\name{distro_req}
\alias{distro_req}
\title{Get system requirements from package name}
\usage{
distro_req(reqs, distro, system_req_folder)
}
\arguments{
\item{reqs}{A \code{data.frame} from \code{\link[=get_package_reqs]{get_package_reqs()}} with free-form system requirements.}

\item{distro}{\verb{[character]} Name of the Linux distribution.}

\item{system_req_folder}{The location of a local clone of "System Requirements for R Packages".}
}
\value{
A \code{data.frame} with columns "Rule" and "Package", where "Package" is the name of systems
package on \code{distro}.
}
\description{
This function use RStudio's repository of system requirements: \url{https://github.com/rstudio/r-system-requirements}.
A local copy/clone of this repository must be available.
}
