% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_deps.R
\name{get_package_deps}
\alias{get_package_deps}
\title{Find package dependencies}
\usage{
get_package_deps(package_name, verbose = FALSE)
}
\arguments{
\item{package_name}{\verb{[character]} Name of the package(s).}

\item{verbose}{\verb{[bool]} Print information along the way.}
}
\value{
A \code{character} vector with names of dependencies -- minus base packages.
}
\description{
Find all non-base packages that a package depends on.
}
